delete functionality

+++ Api/controller/listing.controller.js
@@ -12,13 +12,21 @@ export const createListing = async (req, res, next) => {
 
 export const deleteListing = async (req, res, next) => {
   const listing = await Listing.findById(req.params.id);
+
+  if (!listing) {
+    return next(errorHandler(404, "listing not found!"));
+  }
+
   if (req.user.id !== listing.userRef) {
     return next(errorHandler(401, "You can only delete you own listings!"));
   }
 
   try {
-    await Listing.findByIdAndDelete(req.params.id);
-    res.status(200).json("Listing has been deleted!");
+    const deleted = await Listing.findByIdAndDelete(req.params.id);
+    if (!deleted) {
+      next(errorHandler(404, "failed to delete"));
+    }
+    res.status(200).json({ deleted, message: "Listing has been deleted!" });
   } catch (error) {
     next(error);
   }
